Adding Onboarding logic to an extension
